requirements.txt
fastapi
uvicorn
pydantic
pandas
python-dotenv
openai
chromadb
vanna
psycopg2-binary